+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2022-10-05
+
+### Added
+
+- A secret debug output
+- Light type checking with `// @ts-check`
+
+### Changes
+
+- Don't count time when fewer than two participants
+- Replace instances of thumbnails with participants
+- Expand collapse icons to chevrons
+
 ## [0.0.1] - 2022-09-29
 
 ### Changes